@@ -75,7 +75,7 @@
   "private": true,
   "dependencies": {
     "@angular-devkit/schematics-cli": "^0.12.3",
-    "@angular/animations": "^7.1.3",
+    "@angular/animations": "7.2.1",
     "@angular/cdk": "^7.1.1",
     "@angular/material": "^7.1.1",
     "@ionic/angular": "^4.0.0",
@@ -90,17 +90,17 @@
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.12.2",
-    "@angular/cli": "~7.1.3",
-    "@angular/common": "^7.1.3",
-    "@angular/compiler": "^7.1.3",
-    "@angular/compiler-cli": "^7.1.3",
-    "@angular/core": "^7.1.3",
-    "@angular/forms": "^7.1.3",
-    "@angular/http": "^7.1.3",
+    "@angular/cli": "7.2.1",
+    "@angular/common": "7.2.1",
+    "@angular/compiler": "7.2.1",
+    "@angular/compiler-cli": "7.2.1",
+    "@angular/core": "7.2.1",
+    "@angular/forms": "7.2.1",
+    "@angular/http": "7.2.1",
     "@angular/language-service": "^7.1.3",
-    "@angular/platform-browser": "^7.1.3",
-    "@angular/platform-browser-dynamic": "^7.1.3",
-    "@angular/router": "^7.1.3",
+    "@angular/platform-browser": "7.2.1",
+    "@angular/platform-browser-dynamic": "7.2.1",
+    "@angular/router": "7.2.1",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "~8.9.4",
