@@ -14,11 +14,13 @@ import { LoginComponent } from './components/login/login.component';
 import { JobsComponent } from './components/jobs/jobs.component';
 import { ReleaserComponent } from './components/releaser/releaser.component';
 import { IonicModule } from '@ionic/angular';
+import {JenkinsService} from "./services/jenkins.service";
+import {FormsModule} from "@angular/forms";
 
 @NgModule({
   declarations: [AppComponent, HomeComponent, LoginComponent, JobsComponent, ReleaserComponent],
-  imports: [AppRoutingModule, HttpClientModule, BrowserModule, BrowserAnimationsModule, IonicModule.forRoot()],
-  providers: [ElectronService],
+  imports: [AppRoutingModule, HttpClientModule, BrowserModule, BrowserAnimationsModule, IonicModule.forRoot(), FormsModule],
+  providers: [ElectronService, JenkinsService],
   bootstrap: [AppComponent]
 })
 export class AppModule {}
